LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-url-expected.txt


FAIL A custom property of type <url> can yield a discrete CSS Transition assert_equals: A single animation is running expected 1 but got 0
